P2P fix integrational tests
