+++ elements/lisk-p2p/test/functional/blacklist_whitelist_fixed_peers.ts
@@ -15,12 +15,9 @@
 import { expect } from 'chai';
 import { P2P } from '../../src/index';
 import { wait } from '../utils/helpers';
-import cloneDeep = require('lodash.clonedeep');
-import { SCServerSocket } from 'socketcluster-server';
-import * as url from 'url';
 import { createNetwork, destroyNetwork } from 'utils/network_setup';
 
-describe('Blacklisted/fixed/whitelisted peers', () => {
+describe.only('Blacklisted/fixed/whitelisted peers', () => {
 	const FIVE_CONNECTIONS = 5;
 	const POPULATOR_INTERVAL_WITH_LIMIT = 10;
 	const NETWORK_START_PORT = 5000;
@@ -34,23 +31,6 @@ describe('Blacklisted/fixed/whitelisted peers', () => {
 			number: undefined,
 		},
 	];
-	const serverSocketPrototypeBackup = cloneDeep(SCServerSocket.prototype);
-
-	before(async () => {
-		const serverSocketPrototype = SCServerSocket.prototype as any;
-		const realResetPongTimeoutFunction =
-			serverSocketPrototype._resetPongTimeout;
-		serverSocketPrototype._resetPongTimeout = function() {
-			const queryObject = url.parse(this.request.url, true).query as any;
-			let ipSuffix = queryObject.wsPort - 5000 + 10;
-			this.remoteAddress = `127.0.0.${ipSuffix}`;
-			return realResetPongTimeoutFunction.apply(this, arguments);
-		};
-	});
-
-	after(async () => {
-		SCServerSocket.prototype = serverSocketPrototypeBackup;
-	});
 
 	describe('blacklisting', () => {
 		let p2pNodeList: ReadonlyArray<P2P> = [];
@@ -72,27 +52,10 @@ describe('Blacklisted/fixed/whitelisted peers', () => {
 		];
 
 		beforeEach(async () => {
-			const customSeedPeers = (
-				index: number,
-				startPort: number,
-				networkSize: number,
-			) => [
-				{
-					ipAddress: '127.0.0.' + (((index + 1) % networkSize) + 10),
-					wsPort: startPort + ((index + 1) % networkSize),
-				},
-			];
-
-			const customConfig = (
-				index: number,
-				startPort: number,
-				networkSize: number,
-			) => ({
-				hostIp: '127.0.0.' + (index + 10),
+			const customConfig = () => ({
 				populatorInterval: POPULATOR_INTERVAL_WITH_LIMIT,
 				maxOutboundConnections: FIVE_CONNECTIONS,
 				maxInboundConnections: FIVE_CONNECTIONS,
-				seedPeers: customSeedPeers(index, startPort, networkSize),
 				blacklistedPeers,
 				fixedPeers: blacklistedPeers,
 				whitelistedPeers: blacklistedPeers,
@@ -176,27 +139,10 @@ describe('Blacklisted/fixed/whitelisted peers', () => {
 			},
 		];
 		beforeEach(async () => {
-			const customSeedPeers = (
-				index: number,
-				startPort: number,
-				networkSize: number,
-			) => [
-				{
-					ipAddress: '127.0.0.' + (((index + 1) % networkSize) + 10),
-					wsPort: startPort + ((index + 1) % networkSize),
-				},
-			];
-
-			const customConfig = (
-				index: number,
-				startPort: number,
-				networkSize: number,
-			) => ({
-				hostIp: '127.0.0.' + (index + 10),
+			const customConfig = () => ({
 				populatorInterval: POPULATOR_INTERVAL_WITH_LIMIT,
 				maxOutboundConnections: FIVE_CONNECTIONS,
 				maxInboundConnections: FIVE_CONNECTIONS,
-				seedPeers: customSeedPeers(index, startPort, networkSize),
 				fixedPeers,
 				previousPeers,
 			});
@@ -242,27 +188,10 @@ describe('Blacklisted/fixed/whitelisted peers', () => {
 			},
 		];
 		beforeEach(async () => {
-			const customSeedPeers = (
-				index: number,
-				startPort: number,
-				networkSize: number,
-			) => [
-				{
-					ipAddress: '127.0.0.' + (((index + 1) % networkSize) + 10),
-					wsPort: startPort + ((index + 1) % networkSize),
-				},
-			];
-
-			const customConfig = (
-				index: number,
-				startPort: number,
-				networkSize: number,
-			) => ({
-				hostIp: '127.0.0.' + (index + 10),
+			const customConfig = () => ({
 				populatorInterval: POPULATOR_INTERVAL_WITH_LIMIT,
 				maxOutboundConnections: FIVE_CONNECTIONS,
 				maxInboundConnections: FIVE_CONNECTIONS,
-				seedPeers: customSeedPeers(index, startPort, networkSize),
 				whitelistedPeers,
 				previousPeers,
 			});
