+++ elements/lisk-p2p/test/utils/network_setup.ts
@@ -19,6 +19,9 @@ import {
 	DEFAULT_MAX_OUTBOUND_CONNECTIONS,
 	DEFAULT_MAX_INBOUND_CONNECTIONS,
 } from '../../src';
+import cloneDeep = require('lodash.clonedeep');
+import { SCServerSocket } from 'socketcluster-server';
+import * as url from 'url';
 export const NETWORK_START_PORT = 5000;
 export const NETWORK_PEER_COUNT = 10;
 export const POPULATOR_INTERVAL = 50;
@@ -30,6 +33,8 @@ export const SEED_PEER_IP = '127.0.0.1';
 export const BASE_PEER_IP = '127.0.0.';
 export const NETWORK_CREATION_WAIT_TIME = 1000;
 export const NETWORK_DESTROY_WAIT_TIME = 1000;
+
+const serverSocketPrototypeBackup = cloneDeep(SCServerSocket.prototype);
 
 export const nodeInfoConstants = {
 	nethash: 'da3ed6a45429278bac2666961289ca17ad86595d33b31037615d4b8e8f158bba',
@@ -58,6 +63,15 @@ export const createNetwork = async ({
 	networkDiscoveryWaitTime,
 	customConfig,
 }: TestNetworkConfig = {}) => {
+	const serverSocketPrototype = SCServerSocket.prototype as any;
+	const realResetPongTimeoutFunction = serverSocketPrototype._resetPongTimeout;
+	serverSocketPrototype._resetPongTimeout = function() {
+		const queryObject = url.parse(this.request.url, true).query as any;
+		let ipSuffix = queryObject.wsPort - 5000 + 10;
+		this.remoteAddress = `${BASE_PEER_IP}${ipSuffix}`;
+		return realResetPongTimeoutFunction.apply(this, arguments);
+	};
+
 	const numberOfPeers = networkSize ? networkSize : NETWORK_PEER_COUNT;
 	const startPort = startNodePort ? startNodePort : NETWORK_START_PORT;
 
@@ -68,8 +82,8 @@ export const createNetwork = async ({
 				? []
 				: [
 						{
-							ipAddress: BASE_PEER_IP + index,
-							wsPort: NETWORK_START_PORT + index - 1,
+							ipAddress: BASE_PEER_IP + (((index + 1) % numberOfPeers) + 10),
+							wsPort: startPort + ((index + 1) % numberOfPeers),
 						},
 				  ];
 
@@ -88,7 +102,7 @@ export const createNetwork = async ({
 			populatorInterval: POPULATOR_INTERVAL,
 			maxOutboundConnections: DEFAULT_MAX_OUTBOUND_CONNECTIONS,
 			maxInboundConnections: DEFAULT_MAX_INBOUND_CONNECTIONS,
-			hostIp: BASE_PEER_IP + (index + 1),
+			hostIp: BASE_PEER_IP + (index + 10),
 			nodeInfo: {
 				wsPort: nodePort,
 				nethash: nodeInfoConstants.nethash,
@@ -126,4 +140,6 @@ export const destroyNetwork = async (
 	await wait(
 		networkDestroyWaitTime ? networkDestroyWaitTime : NETWORK_DESTROY_WAIT_TIME,
 	);
+
+	SCServerSocket.prototype = serverSocketPrototypeBackup;
 };
